add functionsにcharacter.pyを追加
